requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
pkg-resources==0.0.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
requests==2.18.4
six==1.10.0
urllib3==1.22
